Renamed module dhancore to dhan_core per convention
